Bug 1102195 Part 1: Update Chromium sandbox code to commit df7cc6c04725630dd4460f29d858a77507343b24. r=aklotz, r=jld
UltraBlame original commit: 38d12afdc7b15e8499a4951f8e7cd0495c600c92

+++ security/sandbox/chromium/sandbox/win/src/crosscall_params.h
@@ -200,14 +200,16 @@ class ActualCallParams : public CrossCallParams {
   
   explicit ActualCallParams(uint32 tag)
       : CrossCallParams(tag, NUMBER_PARAMS) {
-    param_info_[0].offset_ = parameters_ - reinterpret_cast<char*>(this);
+    param_info_[0].offset_ =
+        static_cast<uint32>(parameters_ - reinterpret_cast<char*>(this));
   }
 
   
   
   ActualCallParams(uint32 tag, uint32 number_params)
       : CrossCallParams(tag, number_params) {
-    param_info_[0].offset_ = parameters_ - reinterpret_cast<char*>(this);
+    param_info_[0].offset_ =
+        static_cast<uint32>(parameters_ - reinterpret_cast<char*>(this));
   }
